requirements.txt
Django~=5.1.4
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
qrcode[pil]
PyYAML
drf-yasg
django-rest-authtoken
